Refactoring util into config and connection files
