+++ oneview_redfish_toolkit/conf/redfish.conf
@@ -7,7 +7,7 @@ authentication_mode = session
 debug = false
 
 [oneview_config]
-ip =
+ip = ai-cicvc3-004.vse.rdlabs.hpecorp.net, ai-cicvc3-055.vse.rdlabs.hpecorp.net  
 
 [credentials]
 userName =
@@ -19,8 +19,8 @@ DeliveryRetryIntervalSeconds = 30
 
 [ssl]
 SSLType = self-signed
-SSLCertFile =
-SSLKeyFile =
+SSLCertFile = oneview_redfish_toolkit/certs/self-signed.crt
+SSLKeyFile = oneview_redfish_toolkit/certs/self-signed.key
 
 [ssl-cert-defaults]
 countryName = US
