+++ oneview_redfish_toolkit/tests/test_util.py
@@ -18,18 +18,15 @@
     Tests for store_schema and load_registry function from util.py
 """
 
-import collections
-import configparser
 import json
 import os
 import socket
 
 from oneview_redfish_toolkit.api.errors import OneViewRedfishError
-from oneview_redfish_toolkit.api.errors \
-    import OneViewRedfishResourceNotFoundError
 from oneview_redfish_toolkit.api.event import Event
-from oneview_redfish_toolkit.api import schemas
 from oneview_redfish_toolkit.api.subscription import Subscription
+from oneview_redfish_toolkit import config
+from oneview_redfish_toolkit import connection
 from oneview_redfish_toolkit import util
 import unittest
 from unittest import mock
@@ -46,18 +43,6 @@ class TestUtil(unittest.TestCase):
                 - have all expected options
                 - checks if globals vars are not none
 
-            store_schema()
-                - invalid schema dir
-
-            load_registry()
-                - invalid schema dir
-                - valid schema dir invalid schemas dict
-                - valid schema dir and dict
-
-            get_oneview_client()
-                - connection recover
-                - connection renew
-                - connection failure
     """
 
     # load_conf() tests
@@ -65,171 +50,6 @@ class TestUtil(unittest.TestCase):
         self.schema_dir = './oneview_redfish_toolkit/schemas'
         self.registry_dir = './oneview_redfish_toolkit/registry'
         self.config_file = './oneview_redfish_toolkit/conf/redfish.conf'
-
-    def test_load_conf_invalid_config_file(self):
-        # Tests if passing a file that does not exists returns false.
-        try:
-            util.load_conf('non-exist.conf')
-        except Exception as e:
-            self.assertIsInstance(
-                e,
-                OneViewRedfishResourceNotFoundError
-            )
-
-    def test_load_conf_valid_config_file(self):
-        # Tests if passing a valid file returns a object
-        self.assertIsInstance(util.load_conf(self.config_file),
-                              configparser.ConfigParser)
-
-    def test_load_conf_has_all_expect_sessions(self):
-        # Tests if ini file has all expected sections
-        cfg = util.load_conf(self.config_file)
-
-        self.assertTrue(cfg.has_section('redfish'),
-                        msg='Section {} not found in ini file {}'.format(
-                        'redfish', self.config_file))
-        self.assertTrue(cfg.has_section('oneview_config'),
-                        msg='Section {} not found in ini file {}'.
-                        format('oneview_config', self.config_file))
-        self.assertTrue(cfg.has_section('credentials'),
-                        msg='Section {} not found in ini file {}'.
-                        format('credentials', self.config_file))
-        self.assertTrue(cfg.has_section('event_service'),
-                        msg='Section {} not found in ini file {}'.
-                        format('event_service', self.config_file))
-
-    def test_load_conf_has_all_options(self):
-        # Tests if ini file has all expected options
-
-        cfg = util.load_conf(self.config_file)
-
-        self.assertTrue(cfg.has_option('redfish', 'indent_json'),
-                        msg='Option {} not found in section {} in ini file {}'
-                        .format('indent_json', 'redfish',
-                        self.config_file))
-        self.assertTrue(cfg.has_option('oneview_config', 'ip'),
-                        msg='Option {} not found in section {} in ini file {}'
-                        .format('ip', 'oneview_config', self.config_file))
-        self.assertTrue(cfg.has_option('credentials', 'userName'),
-                        msg='Option {} not found in section {} in ini file {}'
-                        .format('userName', 'credentials',
-                        self.config_file))
-        self.assertTrue(cfg.has_option('credentials', 'password'),
-                        msg='Option {} not found in section {} in ini file {}'
-                        .format('password', 'credentials',
-                        self.config_file))
-
-    # store_schemas() tests
-    def test_store_schemas_invalid_schema_dir(self):
-        # Tests if passing a non existing schema dir returns False
-
-        try:
-            util.store_schemas('non-exist-schema-dir')
-        except Exception as e:
-            self.assertIsInstance(
-                e,
-                OneViewRedfishResourceNotFoundError,
-                msg="Unexpected exception: {}".format(e.msg)
-            )
-
-    # load_registry() tests
-    def test_load_registries_invalid_registry_dir(self):
-        # Tests load_registry() passing a non existing registry dir
-
-        schemas = dict()
-
-        try:
-            util.load_registry('non-exist-registry-dir', schemas)
-        except Exception as e:
-            self.assertIsInstance(
-                e,
-                OneViewRedfishResourceNotFoundError,
-                msg="Unexpected exception: {}".format(e.msg)
-            )
-
-    def test_load_registries_valid_registry_dir_invalid_dict(self):
-        # Tests load_registry() passing a valid registry dir and an invalid
-        # registry dict
-
-        registries = dict()
-        registries['failed'] = 'fail.json'
-
-        try:
-            util.load_registry(self.registry_dir, registries)
-        except Exception as e:
-            self.assertIsInstance(
-                e,
-                OneViewRedfishResourceNotFoundError,
-                msg="Unexpected exception: {}".format(e.msg)
-            )
-
-    def test_load_registries_valid_registry_dir_valid_dict(self):
-        # Tests loading registry files from redfish.conf
-
-        registries = schemas.REGISTRY
-
-        try:
-            registry_dict = util.load_registry(self.registry_dir, registries)
-            self.assertIsInstance(registry_dict, collections.OrderedDict)
-        except Exception as e:
-            self.fail('Failed to load registries files: {}'.format(e.msg))
-
-    @mock.patch.object(util, 'check_oneview_availability')
-    def test_load_config(self, check_ov_availability):
-        # Test load config function
-
-        util.load_config(self.config_file)
-
-        # After running loadconfig all variable should be set
-        self.assertIsNotNone(util.config, msg='Failed do load ini')
-        self.assertIsNotNone(util.ov_config, msg='Failed do create ov_config')
-        self.assertIsNotNone(
-            util.registry_dict, msg='Failed to load registries')
-        self.assertTrue(check_ov_availability.called)
-
-    @mock.patch.object(util, 'OneViewClient')
-    def test_get_ov_client_recover(self, oneview_client_mockup):
-        # Tests a successful recover of a OV client
-
-        oneview_client = oneview_client_mockup()
-        oneview_client.connection.get.return_value = list()
-
-        try:
-            ov_client = util.OneViewClient({})
-        except Exception as e:
-            self.fail('Failed to connect to OneView: '.format(e))
-        self.assertIsNotNone(ov_client)
-
-    @mock.patch.object(util, 'OneViewClient')
-    def test_get_ov_client_renew(self, oneview_client_mockup):
-        """Tests getting a OV client from expired session"""
-
-        oneview_client = oneview_client_mockup()
-        oneview_client.connection.get.return_value = \
-            Exception('session expired')
-        oneview_client.connection.login.return_value = oneview_client
-
-        try:
-            ov_client = util.OneViewClient({})
-        except Exception as e:
-            self.fail('Failed to connect to OneView: '.format(e))
-        self.assertIsNotNone(ov_client)
-
-    @mock.patch.object(util, 'OneViewClient')
-    def test_get_ov_client_oneview_offline(self, oneview_client_mockup):
-        """Tests getting a OV client from an offline oneview"""
-
-        oneview_client = oneview_client_mockup()
-        oneview_client.connection.get.return_value = \
-            Exception('OneView not responding')
-        oneview_client.connection.login.return_value = \
-            Exception('OneView not responding')
-
-        try:
-            ov_client = util.OneViewClient({})
-        except Exception as e:
-            self.fail('Failed to connect to OneView: '.format(e))
-        self.assertIsNotNone(ov_client)
 
     def test_get_ip(self):
         # Tests get_ip function; This test may not work if it returns an IPV6.
@@ -239,12 +59,12 @@ class TestUtil(unittest.TestCase):
         except Exception:
             self.fail("Failed to get a valid IP Address")
 
-    @mock.patch.object(util, 'check_oneview_availability')
+    @mock.patch.object(connection, 'check_oneview_availability')
     def test_create_certs(
         self, check_ov_availability):
         # Test generate_certificate function
 
-        util.load_config(self.config_file)
+        config.load_config(self.config_file)
 
         util.generate_certificate("oneview_redfish_toolkit/certs",
                                   "test", 2048)
@@ -254,13 +74,13 @@ class TestUtil(unittest.TestCase):
         self.assertTrue(os.path.exists(os.path.join("oneview_redfish_toolkit",
                         "certs", "test" + ".key")))
 
-    @mock.patch.object(util, 'check_oneview_availability')
+    @mock.patch.object(connection, 'check_oneview_availability')
     def test_load_event_service_invalid_info(
         self, check_ov_availability):
         self.assertRaises(
-            OneViewRedfishError, util.load_config(self.config_file))
+            OneViewRedfishError, config.load_config(self.config_file))
 
-    @mock.patch.object(util, 'check_oneview_availability')
+    @mock.patch.object(connection, 'check_oneview_availability')
     @mock.patch.object(util, 'subscriptions_by_type')
     @mock.patch(
         'oneview_redfish_toolkit.event_dispatcher.EventDispatcher.start')
@@ -268,7 +88,7 @@ class TestUtil(unittest.TestCase):
         self, start_mock, subscription_mock, check_ov_availability):
         """Tests SubmitTestEvent action with two subscribers"""
 
-        util.load_config(self.config_file)
+        config.load_config(self.config_file)
 
         with open(
             'oneview_redfish_toolkit/mockups/oneview/Alert.json'
@@ -284,7 +104,7 @@ class TestUtil(unittest.TestCase):
 
         self.assertTrue(start_mock.call_count == 2)
 
-    @mock.patch.object(util, 'check_oneview_availability')
+    @mock.patch.object(connection, 'check_oneview_availability')
     @mock.patch.object(util, 'subscriptions_by_type')
     @mock.patch(
         'oneview_redfish_toolkit.event_dispatcher.EventDispatcher.start')
@@ -292,7 +112,7 @@ class TestUtil(unittest.TestCase):
         self, start_mock, subscription_mock, check_ov_availability):
         """Tests SubmitTestEvent action with no subscribers"""
 
-        util.load_config(self.config_file)
+        config.load_config(self.config_file)
 
         with open(
             'oneview_redfish_toolkit/mockups/oneview/Alert.json'
