+++ oneview_redfish_toolkit/util.py
@@ -15,33 +15,24 @@
 # under the License.
 
 # Python libs
-import collections
-import configparser
-import glob
-import json
 import logging
 import logging.config
 import OpenSSL
 import os
 import pkg_resources
 import socket
-import ssl
-import time
-
-# 3rd party libs
-from copy import deepcopy
-from flask_api import status
-from hpOneView.oneview_client import OneViewClient
-from http.client import HTTPSConnection
 
 # Modules own libs
 from oneview_redfish_toolkit.api.errors import OneViewRedfishError
-from oneview_redfish_toolkit.api.errors \
-    import OneViewRedfishResourceNotAccessibleError
-from oneview_redfish_toolkit.api.errors \
-    import OneViewRedfishResourceNotFoundError
-from oneview_redfish_toolkit.api import schemas
+from oneview_redfish_toolkit import config
 from oneview_redfish_toolkit.event_dispatcher import EventDispatcher
+
+
+# Globals vars:
+#   globals()['subscriptions_by_type']
+#   globals()['all_subscriptions']
+#   globals()['delivery_retry_attempts']
+#   globals()['delivery_retry_interval']
 
 globals()['subscriptions_by_type'] = {
     "StatusChange": {},
@@ -56,124 +47,20 @@ globals()['all_subscriptions'] = {}
 API_VERSION = 600
 
 
-def configure_logging(log_file_path):
-    """Loads logging.conf file
-
-        Loads logging.conf file to create the logger configuration.
-
-        The logger configuration has two handlers, one of stream
-        (show logs in the console) and other of file (save a log file)
-        where you can choose one of it in [logger_root : handlers].
-        In it you can choose the logger level as well.
-
-        Level: Numeric value
-        ---------------------
-        CRITICAL: 50
-        ERROR:    40
-        WARNING:  30
-        INFO:     20
-        DEBUG:    10
-        NOTSET:   00
-        ---------------------
-
-        How to use: import logging and logging.exception('message')
-
-        Args:
-            log_file_path: logging.conf path.
-
-        Exception:
-            Exception: if logging.conf file not found.
-    """
-    if os.path.isfile(log_file_path) is False:
-        raise Exception("Config file {} not found".format(log_file_path))
-    else:
-        logging.config.fileConfig(log_file_path)
+def get_subscriptions_by_type():
+    return globals()['subscriptions_by_type']
 
 
-def load_config(conf_file):
-    """Loads redfish.conf file
-
-        Loads and parsers the system conf file into config global var
-        Loads json schemas into schemas_dict global var
-        Established a connection with OneView and sets in as ov_conn
-        global var
-
-        Args:
-            conf_file: string with the conf file name
-
-        Returns:
-            None
-
-        Exception:
-            OneViewRedfishResourceNotFoundError:
-                - if conf file not found
-                - if any of the schemas files are not found
-                - if the schema directory is not found
-            OneViewRedFishResourceNotAccessibleError:
-                - if can't access schema's directory
-            HPOneViewException:
-                - if fails to connect to oneview
-    """
-
-    config = load_conf(conf_file)
-    globals()['config'] = config
-
-    # Config file read set global vars
-    # Setting ov_config
-    ov_config = dict(config.items('oneview_config'))
-    ov_config['credentials'] = dict(config.items('credentials'))
-    ov_config['api_version'] = API_VERSION
-    globals()['ov_config'] = ov_config
-
-    load_event_service_info()
-
-    # Load schemas | Store schemas
-    try:
-        check_oneview_availability(ov_config)
-
-        registry_dict = load_registry(
-            get_registry_path(),
-            schemas.REGISTRY)
-        globals()['registry_dict'] = registry_dict
-
-        store_schemas(get_schemas_path())
-    except OneViewRedfishResourceNotFoundError as e:
-        raise OneViewRedfishError(
-            'Failed to load schemas or registries: {}'.format(e)
-        )
-    except Exception as e:
-        raise OneViewRedfishError(
-            'Failed to connect to OneView: {}'.format(e)
-        )
+def get_all_subscriptions():
+    return globals()['all_subscriptions']
 
 
-def load_conf(conf_file):
-    """Loads and parses conf file
+def get_delivery_retry_attempts():
+    return globals()['delivery_retry_attempts']
 
-        Loads and parses the module conf file
 
-        Args:
-            conf_file: string with the conf file name
-
-        Returns:
-            ConfigParser object with conf_file configs
-
-        Exception:
-            OneViewRedfishResourceNotFoundError:
-                - if conf file not found
-    """
-
-    if not os.path.isfile(conf_file):
-        raise OneViewRedfishResourceNotFoundError(conf_file, 'File')
-
-    config = configparser.ConfigParser()
-    config.optionxform = str
-    try:
-        config.read(conf_file)
-    except Exception:
-        raise
-
-    return config
+def get_delivery_retry_interval():
+    return globals()['delivery_retry_interval']
 
 
 def load_event_service_info():
@@ -186,8 +73,8 @@ def load_event_service_info():
             OneViewRedfishError: DeliveryRetryAttempts and
             DeliveryRetryIntervalSeconds must be integers greater than zero.
     """
-    config = globals()['config']
-    event_service = dict(config.items("event_service"))
+    app_config = config.get_config()
+    event_service = dict(app_config.items("event_service"))
 
     try:
         delivery_retry_attempts = \
@@ -208,160 +95,6 @@ def load_event_service_info():
     globals()['delivery_retry_interval'] = delivery_retry_interval
 
 
-def load_registry(registry_dir, registries):
-    """Loads Registries
-
-        Loads all registries using registry_dir directory
-
-        Args:
-            registry_dir: string with the directory to load registries from
-            registries: dict with registry name as key and registry file_name
-                as value. The key will also be the key in the returning dict.
-
-        Returns:
-            OrderedDict: A dict containing 'RegistryName': registry_obj
-
-        Exceptions:
-            OneviewRedfishResourceNotFoundError:
-                - if registry_dir is not found
-                - any of json files is not found
-            OneviewRedfishResourceNotAccessible:
-                - if registry_dir is can't be accessed
-    """
-
-    if os.path.isdir(registry_dir) is False:
-        raise OneViewRedfishResourceNotFoundError(
-            registry_dir, 'Directory')
-    if os.access(registry_dir, os.R_OK) is False:
-        raise OneViewRedfishResourceNotAccessibleError(
-            registry_dir, 'directory')
-
-    registries_dict = collections.OrderedDict()
-    for key in registries:
-        try:
-            with open(registry_dir + '/' + registries[key]) as f:
-                registries_dict[key] = json.load(f)
-        except Exception:
-            raise OneViewRedfishResourceNotFoundError(
-                registries[key], 'File')
-
-    return registries_dict
-
-
-def store_schemas(schema_dir):
-    """Stores all DMTF JSON Schemas
-
-        Stores all schemas listed in schemas searching schema_dir directory.
-
-        Args:
-            schema_dir: String with the directory to load schemas from.
-
-        Returns:
-            Dictionary: A dict containing ('http://redfish.dmtf.org/schemas/
-                        v1/<schema_file_name>': schema_obj) pairs
-    """
-    schema_paths = glob.glob(schema_dir + '/*.json')
-
-    if not schema_paths:
-        raise OneViewRedfishResourceNotFoundError(
-            "JSON Schemas", "File")
-
-    stored_schemas = dict()
-
-    for path in schema_paths:
-        with open(path) as schema_file:
-            json_schema = json.load(schema_file)
-
-        if os.name == 'nt':
-            file_name = path.split('\\')[-1]
-        else:
-            file_name = path.split('/')[-1]
-        stored_schemas["http://redfish.dmtf.org/schemas/v1/" + file_name] = \
-            json_schema
-
-    globals()['stored_schemas'] = stored_schemas
-
-
-def get_oneview_client(session_id=None, is_service_root=False):
-    """Establishes a OneView connection to be used in the module
-
-        Establishes a OV connection if one does not exists.
-        If one exists, do a single OV access to check if its sill
-        valid. If not tries to establish a new connection.
-        Sets the connection on the ov_conn global var
-
-        Args:
-            session_id: The ID of a valid authenticated session, if the
-            authentication_mode is session. Defaults to None.
-
-            is_service_root: Informs if who is calling this function is the
-            ServiceRoot blueprint. If true, even if authentication_mode is
-            set to session it will use the information on the conf file to
-            return a connection.  This is a workaround to allow ServiceRoot
-            to retrieve the appliance UUID before user logs in.
-
-        Returns:
-            OneViewClient object
-
-        Exceptions:
-            HPOneViewException if can't connect or reconnect to OV
-    """
-
-    config = globals()['config']
-
-    auth_mode = config["redfish"]["authentication_mode"]
-
-    if auth_mode == "conf" or is_service_root:
-        # Doing conf based authentication
-        ov_config = globals()['ov_config']
-        ov_client = None
-
-        # Check if connection is ok yet
-        try:
-            # Check if OneViewClient already exists
-            if 'ov_client' not in globals():
-                globals()['ov_client'] = OneViewClient(ov_config)
-
-            ov_client = globals()['ov_client']
-            ov_client.connection.get('/rest/logindomains')
-            return ov_client
-        # If expired try to make a new connection
-        except Exception:
-            try:
-                logging.exception('Re-authenticated')
-                if ov_client:
-                    ov_client.connection.login(ov_config['credentials'])
-                    return ov_client
-            # if failed abort
-            except Exception:
-                raise
-    else:
-        # Auth mode is session
-        oneview_config = deepcopy(globals()['ov_config'])
-        oneview_config['credentials'] = {"sessionID": session_id}
-        try:
-            oneview_client = OneViewClient(oneview_config)
-            oneview_client.connection.get('/rest/logindomains')
-            return oneview_client
-        except Exception:
-            logging.exception("Failed to recover session based connection")
-            raise
-
-
-def get_ip():
-    """Tries to detect default route IP Address"""
-    s = socket.socket(type=socket.SOCK_DGRAM)
-    try:
-        s.connect(("8.8.8.8", 1))
-        ip = s.getsockname()[0]
-    except Exception as e:
-        logging.exception(e)
-        ip = "127.0.0.1"
-    finally:
-        s.close()
-    return ip
-
-
 def generate_certificate(dir_name, file_name, key_length, key_type="rsa"):
     """Create self-signed cert and key files
 
@@ -377,7 +110,7 @@ def generate_certificate(dir_name, file_name, key_length, key_type="rsa"):
             Raise exceptions on error
     """
 
-    config = globals()['config']
+    app_config = config.get_config()
     private_key = OpenSSL.crypto.PKey()
     if key_type == "rsa":
         private_key.generate_key(OpenSSL.crypto.TYPE_RSA, key_length)
@@ -388,13 +121,13 @@ def generate_certificate(dir_name, file_name, key_length, key_type="rsa"):
         logging.error(message)
         raise OneViewRedfishError(message)
 
-    if not config.has_option("ssl-cert-defaults", "commonName"):
-        config["ssl-cert-defaults"]["commonName"] = get_ip()
+    if not app_config.has_option("ssl-cert-defaults", "commonName"):
+        app_config["ssl-cert-defaults"]["commonName"] = get_ip()
 
     cert = OpenSSL.crypto.X509()
     cert_subject = cert.get_subject()
 
-    cert_defaults = dict(config.items("ssl-cert-defaults"))
+    cert_defaults = dict(app_config.items("ssl-cert-defaults"))
 
     for key, value in cert_defaults.items():
         setattr(cert_subject, key, value)
@@ -413,6 +146,20 @@ def generate_certificate(dir_name, file_name, key_length, key_type="rsa"):
     with open(os.path.join(dir_name, file_name + ".key"), "wt") as f:
         f.write(OpenSSL.crypto.dump_privatekey(
             OpenSSL.crypto.FILETYPE_PEM, private_key).decode("UTF-8"))
+
+
+def get_ip():
+    """Tries to detect default route IP Address"""
+    s = socket.socket(type=socket.SOCK_DGRAM)
+    try:
+        s.connect(("8.8.8.8", 1))
+        ip = s.getsockname()[0]
+    except Exception as e:
+        logging.exception(e)
+        ip = "127.0.0.1"
+    finally:
+        s.close()
+    return ip
 
 
 def dispatch_event(event):
@@ -442,51 +189,6 @@ def dispatch_event(event):
         dispatcher.start()
 
 
-def check_oneview_availability(ov_config):
-    """Check OneView availability by doing a GET request to OneView"""
-    attempts = 3
-    retry_interval_sec = 3
-
-    for attempt_counter in range(attempts):
-        try:
-            connection = HTTPSConnection(
-                ov_config['ip'], context=ssl.SSLContext(ssl.PROTOCOL_TLSv1_2))
-
-            connection.request(
-                method='GET', url='/controller-state.json',
-                headers={'Content-Type': 'application/json'})
-
-            response = connection.getresponse()
-
-            if response.status != status.HTTP_200_OK:
-                message = "OneView is unreachable at {}".format(
-                    ov_config['ip'])
-                raise OneViewRedfishError(message)
-
-            text = response.read().decode('UTF-8')
-            status_ov = json.loads(text)
-
-            if status_ov['state'] != 'OK':
-                message = "OneView state is not OK at {}".format(
-                    ov_config['ip'])
-                raise OneViewRedfishError(message)
-
-            return
-        except Exception as e:
-            logging.exception(
-                'Attempt {} to check OneView availability. '
-                'Error: {}'.format(attempt_counter + 1, e))
-
-            if attempt_counter + 1 < attempts:
-                time.sleep(retry_interval_sec)
-        finally:
-            connection.close()
-
-    message = "After {} attempts OneView is unreachable at {}".format(
-        attempts, ov_config['ip'])
-    raise OneViewRedfishError(message)
-
-
 def get_app_path():
     try:
         source = \
@@ -494,13 +196,3 @@ def get_app_path():
         return source
     except Exception:
         return ""
-
-
-def get_schemas_path():
-    source = get_app_path()
-    return os.path.join(source, "schemas")
-
-
-def get_registry_path():
-    source = get_app_path()
-    return os.path.join(source, "registry")
